chore: change data types

+++ src/generator/MeshF/MeshStructs.hpp
@@ -1,15 +1,15 @@
 #pragma once
 
-struct Coord { // 16 baits
+struct Coord {
   unsigned x;
   unsigned z;
-  double y;
+  float y;
 };
 
-struct Normal { // 24 baits
-  double x;
-  double z;
-  double y;
+struct Normal {
+  float x;
+  float z;
+  float y;
 
   Normal &operator+=(const Normal &other) {
     this->x += other.x;
@@ -20,26 +20,26 @@ struct Normal { // 24 baits
   }
 };
 
-struct UV { // 16 baits
-  double u;
-  double v;
+struct UV {
+  float u;
+  float v;
 };
 
-struct Colour { // 32 baits
-  double RSlope;
-  double GHeight;
-  double BDepth;
-  double aBiome;
+struct Colour {
+  float RSlope;
+  float GHeight;
+  float BDepth;
+  float aBiome;
 };
 
-struct Vertices { // 16 + 24 + 16 + 32 = 88 baits
+struct Vertices {
   Coord coord;
   Normal normal;
   UV uv;
   Colour colour;
 };
 
-struct Indices { // 12 baits
+struct Indices {
   unsigned a;
   unsigned b;
   unsigned c;
